backend/requirements.txt
# Queens Connect Web Chat backend – FastAPI + ADK agent (single venv for backend + queens_connect)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
twilio>=9.0.0
google-adk[extensions]>=0.1.0
litellm>=1.0.0
reportlab>=4.0.0
python-docx>=1.0.0
google-cloud-storage>=2.0.0
